backend/requirements.txt
Django
djangorestframework
django-jazzmin
django-cors-headers
djangorestframework-simplejwt